Make GridColumnElement a fake element (#6)
Firefox, IE and Edge refuse to refer to elements only availabl in the light
DOM. Probable reason is that the elements are not in the actual DOM and
one or another check in the webdrivers conclude that the elements have been
detached. Using such elements always results in a StaleElementReferenceException.